Store rows in time segments
This lays some of the groundwork needed for collapsing rows.

A few big items in this change:
* A timestamp is required on every inserted row
* Rows are grouped into time intervals -- no longer as one big byte array
* Time intervals are represented by N memory mapped files ("segments") of 100MB each.
* The table_file_path now represents a folder, not a file prefix. This
  is because we have many memory mapped files now, so having all of a
  table's files contained within a folder is more tidy.
* Inserted timestamps are truncated to match the duration of the time interval

I'm representing the start of time intervals as ints (seconds since
epoch). Let me know if you think using a proper Time type would be
easier to work with. I think one change we should make regardless is to
name the memory mapped files with readable timestamp strings, rather
than ints.

The persistence test now passes, but the rest of the tests and
benchmarks are broken because their fixtures need to be updated. I'll
tackle that next.