Move query parallelism to be at the interval level
This is based on observation that segment granularity may be too small:
(1) there are a lot of little closures doing the work, (2) the merging
at the end of the query is a non-trivial amount of the query time, and
(3) concurrent queries are hard to control.

This has the nice side effect of getting rid of the gross
"segmentQueryWorker" system. We now create the goroutines close to where
we need them (because we're not creating so many anymore).

Still need to benchmark, tune, and pull constants out into config
params.